requirements.txt
PyPDF2
langchain-groq
python-dotenv
pydantic
langchain
fastapi
python-multipart
uvicorn
python-docx

